Fix gettid usage for earlier than glibc 2.30
The earlier version of glibc didn't support gettid()
wrapper and it causes a build error as undeclared function.
Unfortunately Ubuntu 18.04(bionic) which is a current major
OS destribution, still uses 2.27. For more use cases, it should
be nice to have a macro to help syscall for gettid() in
the thread_pool module with this patch.

Closes: https://github.com/LineairDB/LineairDB/issues/246

ref 1: https://man7.org/linux/man-pages/man2/gettid.2.html
ref 2: https://packages.ubuntu.com/ja/source/bionic/glibc